refactor(spectral tools): AbstractClipView for wave and spectrogram view commons
