+++ src/projectscene/CMakeLists.txt
@@ -142,6 +142,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/viewtrackslistmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/wavepainterproxy.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/wavepainterproxy.h
+    ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/abstractclipview.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/abstractclipview.h
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/spectrogramview.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/spectrogramview.h
     ${CMAKE_CURRENT_LIST_DIR}/view/tracksitemsview/waveview.cpp
