+++ src/projectscene/view/tracksitemsview/waveview.cpp
@@ -32,15 +32,15 @@ static const QColor CLASSIC_SAMPLES_BASE_SELECTED_COLOR = QColor(103, 124, 228);
 static const QColor CLASSIC_RMS_COLOR = QColor(151, 151, 253);                      // Rms: #9797FD
 static const QColor CLASSIC_CLIPPING_COLOR = QColor(239, 71, 111);                  // Clipped: #ef476f
 
-static const float SAMPLE_HEAD_DEFAULT_ALPHA= 0.6;
-static const float SAMPLE_HEAD_CLIP_SELECTED_ALPHA = 0.8;
-static const float SAMPLE_HEAD_DATA_SELECTED_ALPHA = 0.9;
-static const float SAMPLE_STALK_DEFAULT_ALPHA = 0.4;
-static const float SAMPLE_STALK_CLIP_SELECTED_ALPHA = 0.6;
-static const float SAMPLE_STALK_DATA_SELECTED_ALPHA = 0.7;
+static const float SAMPLE_HEAD_DEFAULT_ALPHA = 0.6f;
+static const float SAMPLE_HEAD_CLIP_SELECTED_ALPHA = 0.8f;
+static const float SAMPLE_HEAD_DATA_SELECTED_ALPHA = 0.9f;
+static const float SAMPLE_STALK_DEFAULT_ALPHA = 0.4f;
+static const float SAMPLE_STALK_CLIP_SELECTED_ALPHA = 0.6f;
+static const float SAMPLE_STALK_DATA_SELECTED_ALPHA = 0.7f;
 
 WaveView::WaveView(QQuickItem* parent)
-    : QQuickPaintedItem(parent)
+    : AbstractClipView(parent)
 {
     //! NOTE: Push history state after edit is completed to avoid multiple unecessary calls.
     connect(this, &WaveView::isIsolationModeChanged, [this]() {
@@ -66,14 +66,6 @@ WaveView::WaveView(QQuickItem* parent)
 
 WaveView::~WaveView()
 {
-}
-
-void WaveView::setClipKey(const ClipKey& newClipKey)
-{
-    m_clipKey = newClipKey;
-    emit clipKeyChanged();
-
-    update();
 }
 
 IWavePainter::Params WaveView::getWavePainterParams() const
@@ -179,16 +171,6 @@ void WaveView::paint(QPainter* painter)
     wavePainter()->paint(*painter, m_clipKey.key, params, pType);
 }
 
-ClipKey WaveView::clipKey() const
-{
-    return m_clipKey;
-}
-
-TimelineContext* WaveView::timelineContext() const
-{
-    return m_context;
-}
-
 void WaveView::setTimelineContext(TimelineContext* newContext)
 {
     if (m_context == newContext) {
@@ -229,50 +211,6 @@ void WaveView::setClipColor(const QColor& newClipColor)
     m_clipColor = newClipColor;
     emit clipColorChanged();
 
-    update();
-}
-
-bool WaveView::clipSelected() const
-{
-    return m_clipSelected;
-}
-
-void WaveView::setClipSelected(bool newClipSelected)
-{
-    if (m_clipSelected == newClipSelected) {
-        return;
-    }
-    m_clipSelected = newClipSelected;
-    emit clipSelectedChanged();
-
-    update();
-}
-
-ClipTime WaveView::clipTime() const
-{
-    return m_clipTime;
-}
-
-void WaveView::setClipTime(const ClipTime& newClipTime)
-{
-    if (m_clipTime == newClipTime) {
-        return;
-    }
-    m_clipTime = newClipTime;
-    emit clipTimeChanged();
-
-    update();
-}
-
-double WaveView::channelHeightRatio() const
-{
-    return m_channelHeightRatio;
-}
-
-void WaveView::setChannelHeightRatio(double channelHeightRatio)
-{
-    m_channelHeightRatio = channelHeightRatio;
-    emit channelHeightRatioChanged();
     update();
 }
